+++ src/extension.ts
@@ -7,16 +7,23 @@ import { HoverProvider } from "./hover";
 import { CompletionProvider } from "./completion";
 import { IconLint } from "./lint";
 import { showPreview } from "./preview";
-import { createCompletion } from "./util";
+import {
+  createCompletion,
+  handleDownload,
+  getVersions,
+  IVersionInfo,
+} from "./util";
 
 export function activate(context: vscode.ExtensionContext) {
   const treeDataProvider = new IconTreeDataProvider();
 
+  config.context = context;
+
   const treeView = vscode.window.createTreeView("materialDesignIconsExplorer", {
-    treeDataProvider
+    treeDataProvider,
   });
 
-  treeView.onDidChangeVisibility(event => {
+  treeView.onDidChangeVisibility((event) => {
     if (event.visible) {
       treeDataProvider.refresh();
     }
@@ -37,17 +44,16 @@ export function activate(context: vscode.ExtensionContext) {
   context.subscriptions.push(
     vscode.commands.registerCommand(
       "materialdesigniconsIntellisense.showMdiVersion",
-      () => {
-        fs.readFile(config.mdiPackagePath, (err, data) => {
-          if (err) {
-            vscode.window.showErrorMessage(err.message);
-            return;
-          }
+      async () => {
+        try {
+          const data = await fs.promises.readFile(config.mdiPackagePath);
           vscode.window.showInformationMessage(
             "materialdesignicons-intellisense uses @mdi/svg@" +
               JSON.parse(data.toString("utf8"))["version"]
           );
-        });
+        } catch (err) {
+          vscode.window.showErrorMessage(err.message);
+        }
       }
     )
   );
@@ -82,11 +88,11 @@ export function activate(context: vscode.ExtensionContext) {
         treeView.reveal(
           {
             type: "other",
-            label: "Search results"
+            label: "Search results",
           },
           {
             expand: true,
-            focus: true
+            focus: true,
           }
         );
       }
@@ -101,7 +107,7 @@ export function activate(context: vscode.ExtensionContext) {
           (await vscode.window.showInputBox({
             value: config.lastSearch,
             prompt: "Search icons",
-            placeHolder: "Search icons"
+            placeHolder: "Search icons",
           })) || "";
         vscode.commands.executeCommand(
           "materialdesigniconsIntellisense.performIconSearch",
@@ -113,15 +119,67 @@ export function activate(context: vscode.ExtensionContext) {
 
   context.subscriptions.push(
     vscode.commands.registerCommand(
+      "materialdesigniconsIntellisense.changeMdiVersion",
+      async () => {
+        let items: vscode.QuickPickItem[] | null = null;
+        let info: IVersionInfo | null = null;
+        await vscode.window.withProgress(
+          {
+            location: vscode.ProgressLocation.Notification,
+            cancellable: false,
+          },
+          async (progress) => {
+            progress.report({
+              message: "Getting versions from registry.npmjs.org",
+            });
+            info = await getVersions();
+            await config.updateLatestMdiVersion(info.latest);
+            items = [
+              {
+                label: "latest",
+                description:
+                  `currently ${info.latest}` +
+                  ("latest" === config.rawMdiVersion ? " - selected" : ""),
+              },
+              ...info.versions.map<vscode.QuickPickItem>((v) => ({
+                label: v.version,
+                description:
+                  v.time +
+                  (v.version === config.rawMdiVersion ? " - selected" : ""),
+              })),
+            ];
+          }
+        );
+        const result = await vscode.window.showQuickPick<vscode.QuickPickItem>(
+          items!,
+          {
+            canPickMany: false,
+            placeHolder: `Current version: ${config.rawMdiVersion}`,
+          }
+        );
+        if (result) {
+          const selectedVersion = result.label;
+          const versionToDownload =
+            selectedVersion === "latest" ? info!.latest : selectedVersion;
+          await handleDownload(versionToDownload, info!);
+          await config.updateMdiVersion(selectedVersion);
+          treeDataProvider.refresh();
+        }
+      }
+    )
+  );
+
+  context.subscriptions.push(
+    vscode.commands.registerCommand(
       "materialdesigniconsIntellisense.changeInsertStyle",
       async () => {
         const result = (await vscode.window.showQuickPick(
           ["kebabCase", "camelCase", "homeAssistant"].filter(
-            t => t !== config.insertType
+            (t) => t !== config.insertType
           ),
           {
             canPickMany: false,
-            placeHolder: `Currently Selected: ${config.insertType}`
+            placeHolder: `Currently selected: ${config.insertType}`,
           }
         )) as CompletionType | undefined;
         if (result) {
@@ -134,11 +192,51 @@ export function activate(context: vscode.ExtensionContext) {
 
   context.subscriptions.push(
     vscode.commands.registerCommand(
+      "materialdesigniconsIntellisense.changeLanguages",
+      async () => {
+        const languages = await vscode.languages.getLanguages();
+        const selected = config.selector;
+
+        const selectedButNotAvailable: string[] = [];
+        for (const s of selected) {
+          if (!languages.includes(s)) {
+            selectedButNotAvailable.push(s);
+          }
+        }
+        const items = languages.map<vscode.QuickPickItem>((l) => ({
+          label: l,
+          picked: selected.includes(l),
+        }));
+        items.push(
+          ...selectedButNotAvailable.map<vscode.QuickPickItem>((l) => ({
+            label: l,
+            picked: true,
+            description: "This language is currently not installed",
+          }))
+        );
+
+        const result = await vscode.window.showQuickPick<vscode.QuickPickItem>(
+          items,
+          {
+            canPickMany: true,
+            matchOnDescription: true,
+            matchOnDetail: true,
+          }
+        );
+        if (result) {
+          await config.updateSelector(result.map((r) => r.label));
+        }
+      }
+    )
+  );
+
+  context.subscriptions.push(
+    vscode.commands.registerCommand(
       "materialdesigniconsIntellisense.performIconSearch",
       (search?: string) => {
         if (!search) {
           return vscode.window.showInformationMessage(
-            "Use the MDI Explorer view to search icons"
+            "Use the MDI explorer view to search icons"
           );
         }
         config.lastSearch = search;
@@ -147,11 +245,11 @@ export function activate(context: vscode.ExtensionContext) {
         treeView.reveal(
           {
             type: "other",
-            label: "Search results"
+            label: "Search results",
           },
           {
             expand: true,
-            focus: true
+            focus: true,
           }
         );
       }
@@ -210,7 +308,7 @@ export function activate(context: vscode.ExtensionContext) {
   }
 
   context.subscriptions.push(
-    vscode.workspace.onDidChangeConfiguration(event => {
+    vscode.workspace.onDidChangeConfiguration((event) => {
       if (
         event.affectsConfiguration(
           "materialdesigniconsIntellisense.enableLinter"
@@ -228,7 +326,7 @@ export function activate(context: vscode.ExtensionContext) {
   );
 
   context.subscriptions.push(
-    vscode.workspace.onDidChangeConfiguration(event => {
+    vscode.workspace.onDidChangeConfiguration((event) => {
       if (
         event.affectsConfiguration("materialdesigniconsIntellisense.selector")
       ) {
@@ -239,12 +337,25 @@ export function activate(context: vscode.ExtensionContext) {
       if (
         event.affectsConfiguration(
           "materialdesigniconsIntellisense.overrideFontPackagePath"
-        )
+        ) ||
+        event.affectsConfiguration("materialdesigniconsIntellisense.mdiVersion")
       ) {
         treeDataProvider.refresh();
       }
     })
   );
+
+  // auto update
+  if (config.rawMdiVersion === "latest") {
+    (async () => {
+      const info = await getVersions();
+      if (config.mdiVersion !== info.latest) {
+        await handleDownload(info.latest, info);
+        await config.updateLatestMdiVersion(info.latest);
+        treeDataProvider.refresh();
+      }
+    })();
+  }
 
   console.log('"materialdesignicons-intellisense" is now active');
 }